+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 ### Removed
 - Java 8-10 support, by @HardNorth
 
+## [5.3.15]
+### Fixed
+- Ensure all logs are sent before JVM exit, by @HardNorth
+- Issue [#293](https://github.com/reportportal/client-java/issues/293): Minimum IO pool size, by @HardNorth
+- Issue [#292](https://github.com/reportportal/client-java/issues/292): Truncate too long string fields, by @HardNorth
+
 ## [5.3.14]
 ### Added
 - More evaluation cases in `StatusEvaluation.evaluateStatus` utility method, by @HardNorth